back/requirements.txt
Flask==3.0.3
requests==2.32.3
jsonify==0.5
pillow==10.4.0
torch==2.4.1
diffusers==0.30.3
